refactor: decompose coordinator schedule helpers
